+++ packages/isolate-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ricsam/isolate-client
 
+## 0.1.14
+
+### Patch Changes
+
+- remove baseUrl config option from playwright, that should be configured form the pw instance
+- Updated dependencies
+  - @ricsam/isolate-protocol@0.1.12
+  - @ricsam/isolate-playwright@0.1.12
+
 ## 0.1.13
 
 ### Patch Changes